app/requirements.txt
alchemy==20.5
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.2.1
certifi==2025.1.31
chardet==3.0.4
click==8.1.8
docutils==0.21.2
ecdsa==0.19.0
fastapi==0.115.8
filelock==3.0.12
greenlet==3.1.1
h11==0.14.0
idna==2.8
jose==1.0.0
lockfile==0.12.2
loguru==0.7.3
Mako==1.3.9
MarkupSafe==3.0.2
passlib==1.7.4
pillow==11.1.0
pyasn1==0.6.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-daemon==2.2.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.22.0
rsa==4.9
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
urllib3==1.25.11
uvicorn==0.34.0
